project/frontend/requirements.txt
Flask
requests
python-dotenv



